refactor: 승부예측 카드 리스트를 컴포넌트로 분리
